system_state after rebase
